sidaku/requirementsidaku.txt
appconf==0.1.11
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==4.0.0
django-cities==0.6
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-import-export==3.0.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-sort-order-field==1.2
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
import-export==0.2.67.dev6
MarkupPy==1.14
mysql-connector-python==8.0.30
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
protobuf==3.20.1
PyYAML==5.1.2
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tablib==3.2.1
tqdm==4.64.0
xlrd==2.0.1
xlwt==1.3.0
